Fixed unit tests

+++ test/no/flatline/HuffmanTest.java
@@ -1,13 +1,11 @@
 package no.flatline;
 
 import no.flatline.file.FileUtil;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
 import java.io.*;
 import java.nio.file.FileSystems;
-import java.nio.file.Files;
 import java.nio.file.Path;
 
 import static org.junit.Assert.assertEquals;
@@ -20,10 +18,9 @@ import static org.junit.Assert.assertTrue;
  */
 public class HuffmanTest {
 
-    private static final Path DIRECTORY = FileSystems.getDefault().getPath("resource", "testFiles");
+    private static final Path DIRECTORY = FileSystems.getDefault().getPath("resource/source");
     private static File FILE_1;
     private static File FILE_2;
-    private static File FILE_3;
 
     /**
      * Clears test file directory before initializing class.
@@ -33,12 +30,39 @@ public class HuffmanTest {
     public static void initialSetUp() throws Exception {
         System.out.println("Running initial setup for Huffman Tests.");
         File[] testFiles = DIRECTORY.toFile().listFiles();
+        delete(testFiles);
+        File[] compFiles = FileSystems.getDefault().getPath("resource\\compressed").toFile().listFiles();
+        delete(compFiles);
+        File[] decompFiles = FileSystems.getDefault().getPath("resource\\decompressed").toFile().listFiles();
+        delete(decompFiles);
+        System.out.println("\nRunning setup for a test.");
+        FILE_1 = FileUtil.createFile("resource\\source", "test-1", "txt");
+        FILE_2 = FileUtil.createFile("resource\\source", "test-2", "txt");
+//        FILE_3 = Files.createTempFile(DIRECTORY, "test-3", ".txt").toFile();
+        BufferedWriter bw1 = new BufferedWriter(new FileWriter(FILE_1));
+        BufferedWriter bw2 = new BufferedWriter(new FileWriter(FILE_2));
+//        BufferedWriter bw3 = new BufferedWriter(new FileWriter(FILE_3));
+        bw1.write("Dette er en test for komprimeringen av filer.\nDette er ekstra tekst for å gjøre filstørrelsen stø");
+        bw1.close();
+        System.out.println("INFO:\tGenerated test file " + FILE_1.getName());
+        bw2.write("                                                                                                                                   \u0007                                                       \u0001                                                                                       \u0001                                                                                                                   \u0001               \b   \u0002   \u0001       \u0003       \u0001   \u0001   \u0002   \u0003   \u0002   \u0001       \u0005   \u0001   \u0004       \u0001                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    ");
+        bw2.close();
+        System.out.println("INFO:\tGenerated test file " + FILE_2.getName());
+//        bw3.close();
+//        System.out.println("INFO:\tGenerated test file " + FILE_3.getName());
+
+    }
+
+    private static void delete(File[] testFiles) {
         if (testFiles != null) {
-            int counter = 0;
+            int counter = testFiles.length;
             for (File file : testFiles) {
-                if (file.delete()) {
-                    System.out.println("INFO:\tDeleting test file " + file.getName());
-                    counter++;
+                if (file.getName().startsWith("test")) {
+                    if (file.delete()) {
+                        System.out.println("INFO:\tDeleting test file " + file.getName());
+                    } else {
+                        counter = 0;
+                    }
                 }
             }
             if (counter != testFiles.length) {
@@ -47,52 +71,28 @@ public class HuffmanTest {
         }
     }
 
-    /**
-     * Generate temporary files to use within the tests.
-     * @throws Exception - Throws any exception.
-     */
-    @Before
-    public void setUp() throws Exception {
-        System.out.println("\nRunning setup for a test.");
-        FILE_1 = Files.createTempFile(DIRECTORY, "test-1-", ".txt").toFile();
-        FILE_2 = Files.createTempFile(DIRECTORY, "test-2-", ".txt").toFile();
-        FILE_3 = Files.createTempFile(DIRECTORY, "test-3-", ".txt").toFile();
-        BufferedWriter bw1 = new BufferedWriter(new FileWriter(FILE_1));
-        BufferedWriter bw2 = new BufferedWriter(new FileWriter(FILE_2));
-        BufferedWriter bw3 = new BufferedWriter(new FileWriter(FILE_3));
-        bw1.write("Dette er en test for komprimeringen av filer.");
-        bw1.close();
-        System.out.println("INFO:\tGenerated test file " + FILE_1.getName());
-        bw2.write("");
-        bw2.close();
-        System.out.println("INFO:\tGenerated test file " + FILE_2.getName());
-        bw3.write("                                                                                                                                   \u0007                                                       \u0001                                                                                       \u0001                                                                                                                   \u0001               \b   \u0002   \u0001       \u0003       \u0001   \u0001   \u0002   \u0003   \u0002   \u0001       \u0005   \u0001   \u0004       \u0001                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    ");
-        bw3.close();
-        System.out.println("INFO:\tGenerated test file " + FILE_3.getName());
-    }
-
 
     @Test
     public void compress() throws Exception {
         Huffman h = new Huffman();
         long size = FILE_1.length();
         h.compress(FILE_1);
-        String ex = FileUtil.getExtension(FILE_1);
-        File newFile = new File(FILE_1.getPath().replace(ex, "v2" + ex));
-        assertTrue(FILE_1.length() < size);
+        File newFile = new File(FILE_1.toString().replace("resource\\source", "resource\\compressed") + ".cff");
+        assertTrue(newFile.length() < size);
     }
 
     @Test
     public void decompress() throws Exception {
         Huffman h = new Huffman();
-        // TODO create copy of file to test for equality
-        File f = null;
-        h.compress(FILE_1);
-        h.decompress(FILE_1);
 
-        DataInputStream dis = new DataInputStream(new BufferedInputStream(new FileInputStream(FILE_1)));
-        DataInputStream dat = new DataInputStream(new BufferedInputStream(new FileInputStream(FILE_2)));
+        h.compress(FILE_2);
+        h.decompress(new File(FILE_2.toString().replace("resource\\source", "resource\\compressed") + ".cff"));
 
+        File decomp = new File(FILE_2.toString().replace("resource\\source", "resource\\decompressed"));
+        DataInputStream dis = new DataInputStream(new BufferedInputStream(new FileInputStream(FILE_2)));
+        DataInputStream dat = new DataInputStream(new BufferedInputStream(new FileInputStream(decomp)));
+
+        assertEquals(dis.available(), dat.available());
         int c;
         while ((c = dis.read()) != -1) {
             assertEquals(c, dat.read());
